Added first version of mobile design for company application page
